Rename versioned module and drop Client class
The client class just creates a httpx.Client instance. Therefore it
should be a function. Rename the module and remove the v from the name
to be consistent with other module names.
